move ContestedIssues, addIncludedIssues and createIssueName out of 995 and 996, alphabetize and update imports
