+++ src/applications/appeals/995/tests/utils/submit.unit.spec.js
@@ -1,27 +1,29 @@
 import { expect } from 'chai';
 
 import {
-  PRIMARY_PHONE,
-  EVIDENCE_VA,
   EVIDENCE_OTHER,
   EVIDENCE_PRIVATE,
+  EVIDENCE_VA,
+  PRIMARY_PHONE,
 } from '../../constants';
 import { getDate } from '../../utils/dates';
 import {
-  getTimeZone,
-  createIssueName,
-  getContestedIssues,
-  addIncludedIssues,
   getAddress,
-  getPhone,
-  hasDuplicateLocation,
   getEvidence,
-  hasDuplicateFacility,
   getForm4142,
+  getPhone,
+  getTimeZone,
+  hasDuplicateFacility,
+  hasDuplicateLocation,
 } from '../../utils/submit';
 
 import { SELECTED } from '../../../shared/constants';
-import { removeEmptyEntries } from '../../../shared/utils/submit';
+import {
+  addIncludedIssues,
+  createIssueName,
+  getContestedIssues,
+  removeEmptyEntries,
+} from '../../../shared/utils/submit';
 
 const text =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit, seddo eiusmod tempor incididunt ut labore et dolore magna aliqua. Utenim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.';
