+++ src/applications/appeals/996/config/submit-transformer.js
@@ -1,15 +1,16 @@
 import { DEFAULT_BENEFIT_TYPE } from '../constants';
 
 import {
-  getRep,
-  getConferenceTime, // v2
-  addIncludedIssues,
   addAreaOfDisagreement,
-  getContact,
   getAddress,
+  getConferenceTime, // v2
+  getContact,
   getPhone,
+  getRep,
   getTimeZone,
 } from '../utils/submit';
+
+import { addIncludedIssues } from '../../shared/utils/submit';
 
 export function transform(formConfig, form) {
   // https://dev-developer.va.gov/explore/appeals/docs/decision_reviews?version=current
