+++ src/applications/appeals/996/utils/submit.js
@@ -1,16 +1,8 @@
 import { CONFERENCE_TIMES_V2 } from '../constants';
 
-import {
-  MAX_LENGTH,
-  SELECTED,
-  SUBMITTED_DISAGREEMENTS,
-} from '../../shared/constants';
+import { MAX_LENGTH, SUBMITTED_DISAGREEMENTS } from '../../shared/constants';
 import '../../shared/definitions';
-import { returnUniqueIssues } from '../../shared/utils/issues';
-import {
-  replaceSubmittedData,
-  fixDateFormat,
-} from '../../shared/utils/replace';
+import { replaceSubmittedData } from '../../shared/utils/replace';
 import { removeEmptyEntries } from '../../shared/utils/submit';
 
 // We require the user to input a 10-digit number; assuming we get a 3-digit
@@ -56,96 +48,6 @@ export const getTimeZone = () =>
   // supports IE11
   // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat/resolvedOptions
   Intl.DateTimeFormat().resolvedOptions().timeZone;
-
-/**
- * Combine issues values into one field
- * @param {ContestableIssueAttributes} attributes
- * @returns {String} Issue name - rating % - description combined
- */
-export const createIssueName = ({ attributes } = {}) => {
-  const {
-    ratingIssueSubjectText,
-    ratingIssuePercentNumber,
-    description,
-  } = attributes;
-  const result = [
-    ratingIssueSubjectText,
-    `${ratingIssuePercentNumber || '0'}%`,
-    description,
-  ]
-    .filter(part => part)
-    .join(' - ');
-  return replaceSubmittedData(result).substring(0, MAX_LENGTH.ISSUE_NAME);
-};
-
-/* submitted contested issue format
-[{
-  "type": "contestableIssue",
-  "attributes": {
-    "issue": "tinnitus - 10% - some longer description",
-    "decisionDate": "1900-01-01",
-    "decisionIssueId": 1,
-    "ratingIssueReferenceId": "2",
-    "ratingDecisionReferenceId": "3",
-    "socDate": "2000-01-01"
-  }
-}]
-*/
-export const getContestedIssues = ({ contestedIssues } = {}) =>
-  (contestedIssues || []).filter(issue => issue[SELECTED]).map(issue => {
-    const attr = issue.attributes;
-    const attributes = [
-      'decisionIssueId',
-      'ratingIssueReferenceId',
-      'ratingDecisionReferenceId',
-      'socDate',
-    ].reduce(
-      (acc, key) => {
-        // Don't submit null or empty strings
-        if (attr[key]) {
-          acc[key] = attr[key];
-        }
-        return acc;
-      },
-      {
-        issue: createIssueName(issue),
-        decisionDate: fixDateFormat(attr.approxDecisionDate),
-      },
-    );
-
-    return {
-      // type: "contestableIssues"
-      type: issue.type,
-      attributes,
-    };
-  });
-
-/**
- * Combine included issues and additional issues
- * @param {FormData}
- * @returns {ContestableIssueSubmittable}
- */
-export const addIncludedIssues = formData => {
-  const issues = getContestedIssues(formData);
-
-  const result = issues.concat(
-    (formData.additionalIssues || []).reduce((issuesToAdd, issue) => {
-      if (issue[SELECTED] && issue.issue && issue.decisionDate) {
-        // match contested issue pattern
-        issuesToAdd.push({
-          type: 'contestableIssue',
-          attributes: {
-            issue: replaceSubmittedData(issue.issue),
-            decisionDate: fixDateFormat(issue.decisionDate),
-          },
-        });
-      }
-      return issuesToAdd;
-    }, []),
-  );
-  // Ensure only unique entries are submitted
-  return returnUniqueIssues(result);
-};
 
 /**
  * Add area of disagreement
