+++ src/applications/appeals/shared/utils/submit.js
@@ -1,5 +1,6 @@
-import { MAX_LENGTH, SUBMITTED_DISAGREEMENTS } from '../constants';
-import { replaceSubmittedData } from './replace';
+import { MAX_LENGTH, SELECTED, SUBMITTED_DISAGREEMENTS } from '../constants';
+import { fixDateFormat, replaceSubmittedData } from './replace';
+import { returnUniqueIssues } from './issues';
 
 /**
  * Remove objects with empty string values; Lighthouse doesn't like `null`
@@ -11,6 +12,112 @@ export const removeEmptyEntries = object =>
   Object.fromEntries(
     Object.entries(object).filter(([_, value]) => value !== ''),
   );
+
+/**
+ * Combine issues values into one field
+ * @param {ContestableIssueAttributes} attributes
+ * @returns {String} Issue name - rating % - description combined
+ */
+export const createIssueName = ({ attributes } = {}) => {
+  const {
+    ratingIssueSubjectText,
+    ratingIssuePercentNumber,
+    description,
+  } = attributes;
+  const result = [
+    ratingIssueSubjectText,
+    `${ratingIssuePercentNumber || '0'}%`,
+    description,
+  ]
+    .filter(part => part)
+    .join(' - ');
+  return replaceSubmittedData(result).substring(0, MAX_LENGTH.ISSUE_NAME);
+};
+
+/* submitted contested issue format
+[{
+  "type": "contestableIssue",
+  "attributes": {
+    "issue": "tinnitus - 10% - some longer description",
+    "decisionDate": "1900-01-01",
+    "decisionIssueId": 1,
+    "ratingIssueReferenceId": "2",
+    "ratingDecisionReferenceId": "3",
+    "socDate": "2000-01-01"
+  }
+}]
+*/
+export const getContestedIssues = ({ contestedIssues } = {}) =>
+  (contestedIssues || []).filter(issue => issue[SELECTED]).map(issue => {
+    const attr = issue.attributes;
+    const attributes = [
+      'decisionIssueId',
+      'ratingIssueReferenceId',
+      'ratingDecisionReferenceId',
+      'socDate',
+    ].reduce(
+      (acc, key) => {
+        // Don't submit null or empty strings
+        if (attr[key]) {
+          acc[key] = attr[key];
+        }
+        return acc;
+      },
+      {
+        issue: createIssueName(issue),
+        decisionDate: fixDateFormat(attr.approxDecisionDate),
+      },
+    );
+
+    return {
+      // type: "contestableIssues"
+      type: issue.type,
+      attributes,
+    };
+  });
+
+/**
+ * Combine included issues and additional issues
+ * @param {FormData}
+ * @returns {ContestableIssueSubmittable}
+ */
+export const addIncludedIssues = formData => {
+  const issues = getContestedIssues(formData);
+
+  const result = issues.concat(
+    (formData.additionalIssues || []).reduce((issuesToAdd, issue) => {
+      if (issue[SELECTED] && issue.issue && issue.decisionDate) {
+        // match contested issue pattern
+        issuesToAdd.push({
+          type: 'contestableIssue',
+          attributes: {
+            issue: replaceSubmittedData(issue.issue),
+            decisionDate: fixDateFormat(issue.decisionDate),
+          },
+        });
+      }
+      return issuesToAdd;
+    }, []),
+  );
+  // Ensure only unique entries are submitted
+  return returnUniqueIssues(result);
+};
+
+/**
+ * Strip out extra profile phone data
+ * @param {Veteran} veteran - Veteran formData object
+ * @returns {Object} submittable address
+ */
+export const getPhone = ({ veteran = {} } = {}) => {
+  const truncate = (value, max) =>
+    replaceSubmittedData(veteran.phone?.[value] || '').substring(0, max);
+  return removeEmptyEntries({
+    countryCode: truncate('countryCode', MAX_LENGTH.PHONE_COUNTRY_CODE),
+    areaCode: truncate('areaCode', MAX_LENGTH.PHONE_AREA_CODE),
+    phoneNumber: truncate('phoneNumber', MAX_LENGTH.PHONE_NUMBER),
+    phoneNumberExt: truncate('phoneNumberExt', MAX_LENGTH.PHONE_NUMBER_EXT),
+  });
+};
 
 /**
  * Add area of disagreement
